feat: Adding a new abstract class as base class for form question types
